refactor(localStorage): 🔧 Mueve funciones de localStorage a un servicio dedicado
- Crea nuevo archivo src/services/localStorageService.js para manejar operaciones de localStorage
- Elimina funciones relacionadas con localStorage de src/variables/store.js
- Actualiza importaciones en App.vue, ContentForm.vue y RecoveryModal.vue para usar el nuevo servicio
- Mantiene src/variables/store.js exclusivamente para variables exportadas

Este cambio mejora la organización del código, separando la lógica de almacenamiento
de la gestión de estado global, y facilita el mantenimiento y las pruebas futuras.
